Player Lives + Game Over popup
The player lives system is finally in place and in working order. I also made it so the game over feature works so when the player reaches 0 lives, the game will display a Game Over sign (I still need to freeze the game state and add some more UI functionality but I think im on decent pace.